Formatting; handling of target parameters simplified
I found the creation of star_output confusing and I tried to clean it.
I think mixing parameters from hip, simbad, and user should not happen inside barycorrpy.
An exception might be the rv. which is usually not provided by astrometry. RV is not given by hip and can also miss in simbad.
Therefore, the user may want to overwrite this.
Now this is handled  explicitely,